PIM-3594: move query filter operators
