+++ src/Pim/Bundle/CatalogBundle/Doctrine/Common/Filter/CategoryFilter.php
@@ -6,7 +6,7 @@ use Doctrine\ORM\QueryBuilder;
 use Pim\Bundle\CatalogBundle\Doctrine\Common\ObjectIdResolverInterface;
 use Pim\Bundle\CatalogBundle\Query\FieldFilterHelper;
 use Pim\Bundle\CatalogBundle\Query\FieldFilterInterface;
-use Pim\Bundle\CatalogBundle\Query\Operators;
+use Pim\Bundle\CatalogBundle\Query\Filter\Operators;
 use Pim\Bundle\CatalogBundle\Entity\Repository\CategoryRepository;
 use Pim\Bundle\CatalogBundle\Repository\ProductCategoryRepositoryInterface;
