+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Filter/GroupsFilter.php
@@ -5,7 +5,7 @@ namespace Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter;
 use Pim\Bundle\CatalogBundle\Doctrine\Common\ObjectIdResolverInterface;
 use Pim\Bundle\CatalogBundle\Query\FieldFilterHelper;
 use Pim\Bundle\CatalogBundle\Query\FieldFilterInterface;
-use Pim\Bundle\CatalogBundle\Query\Operators;
+use Pim\Bundle\CatalogBundle\Query\Filter\Operators;
 
 /**
  * Entity filter
