+++ src/Pim/Bundle/CatalogBundle/Query/Filter/Operators.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\Query;
+namespace Pim\Bundle\CatalogBundle\Query\Filter;
 
 /**
  * Filter operators
