+++ src/Pim/Bundle/FilterBundle/Filter/Product/FamilyFilter.php
@@ -4,7 +4,7 @@ namespace Pim\Bundle\FilterBundle\Filter\Product;
 
 use Oro\Bundle\FilterBundle\Datasource\FilterDatasourceAdapterInterface;
 use Oro\Bundle\FilterBundle\Filter\FilterUtility;
-use Pim\Bundle\CatalogBundle\Query\Operators;
+use Pim\Bundle\CatalogBundle\Query\Filter\Operators;
 use Pim\Bundle\FilterBundle\Filter\AjaxChoiceFilter;
 
 /**
